+++ django_typer/tests/apps/test_app/management/commands/adapted.py
@@ -1,0 +1,9 @@
+import json
+
+
+from django_typer import TyperCommand
+
+
+class Command(TyperCommand):
+    def handle(self, message: str):
+        self.echo(f"test_app: {message}")