+++ django_typer/tests/settings/adapted1.py
@@ -1,6 +1,7 @@
 from .base import *
 
 INSTALLED_APPS = [
+    "django_typer.tests.apps.adapter1",
     "django_typer.tests.apps.test_app2",
     *INSTALLED_APPS,
 ]